test: organize integration tests by command
Replace rollout-phase buckets with stable command-focused targets and centralize shared setup without changing coverage.
