+++ tests/tag.rs
@@ -1,175 +1,14 @@
-//! End-to-end tests for the tag/publish layer, using a fake gleam binary
-//! (TRELLIS_GLEAM_BIN), a mock GitHub API (TRELLIS_GITHUB_API_URL), a mock
-//! Hex API served from a local thread (TRELLIS_HEX_API_URL), and real git
-//! repos.
+//! End-to-end tests for `trellis tag`.
 
 mod common;
 
-use assert_cmd::Command;
+use common::{
+    bare_origin, commit_of, copy_fixture_to, git, git_stdout, init_repo, series_repo, set_version,
+    trellis_github, trellis_with_local_http as trellis, write,
+};
 use predicates::prelude::*;
 use std::fs;
-use std::io::{Read, Write as IoWrite};
-use std::net::TcpListener;
 use std::os::unix::fs::PermissionsExt;
-use std::path::{Path, PathBuf};
-
-fn fixture(name: &str) -> PathBuf {
-    Path::new(env!("CARGO_MANIFEST_DIR"))
-        .join("tests/fixtures")
-        .join(name)
-}
-
-/// The mock Hex server binds localhost, so the agent proxy configured in
-/// some environments must not intercept requests.
-fn trellis(dir: &Path) -> Command {
-    let mut cmd = Command::cargo_bin("trellis").unwrap();
-    cmd.current_dir(dir);
-    for var in [
-        "HTTP_PROXY",
-        "HTTPS_PROXY",
-        "http_proxy",
-        "https_proxy",
-        "ALL_PROXY",
-        "all_proxy",
-    ] {
-        cmd.env_remove(var);
-    }
-    cmd
-}
-
-fn write(path: &Path, content: &str) {
-    fs::create_dir_all(path.parent().unwrap()).unwrap();
-    fs::write(path, content).unwrap();
-}
-
-fn copy_fixture_to(root: &Path) {
-    fn walk(dir: &Path, files: &mut Vec<PathBuf>) {
-        for entry in fs::read_dir(dir).unwrap() {
-            let path = entry.unwrap().path();
-            if path.is_dir() {
-                walk(&path, files);
-            } else {
-                files.push(path);
-            }
-        }
-    }
-    let from = fixture("basic");
-    let mut files = Vec::new();
-    walk(&from, &mut files);
-    for file in files {
-        let dest = root.join(file.strip_prefix(&from).unwrap());
-        fs::create_dir_all(dest.parent().unwrap()).unwrap();
-        fs::copy(&file, &dest).unwrap();
-    }
-}
-
-fn git(root: &Path, args: &[&str]) {
-    let status = std::process::Command::new("git")
-        .args(args)
-        .current_dir(root)
-        .env("GIT_AUTHOR_NAME", "t")
-        .env("GIT_AUTHOR_EMAIL", "t@t")
-        .env("GIT_COMMITTER_NAME", "t")
-        .env("GIT_COMMITTER_EMAIL", "t@t")
-        .stdout(std::process::Stdio::null())
-        .stderr(std::process::Stdio::null())
-        .status()
-        .unwrap();
-    assert!(status.success(), "git {args:?} failed");
-}
-
-fn init_repo(root: &Path) {
-    git(root, &["init", "-q", "-b", "main"]);
-    git(root, &["add", "."]);
-    git(root, &["commit", "-q", "-m", "init"]);
-}
-
-/// A fake gleam that logs every invocation (cwd + args) to `.fake/gleam-log`
-/// and snapshots gleam.toml at publish time so tests can observe the rewrite.
-///
-/// It also records whether `build/packages` existed at each invocation, in
-/// `.fake/build-state` — a separate file so tests asserting on the exact
-/// contents of `gleam-log` are unaffected.
-fn install_fake_gleam(root: &Path) -> PathBuf {
-    let script = root.join("fake-gleam.sh");
-    write(
-        &script,
-        &format!(
-            concat!(
-                "#!/bin/sh\n",
-                "set -eu\n",
-                "root=\"{root}\"\n",
-                "echo \"$(basename \"$PWD\") gleam $*\" >> \"$root/.fake/gleam-log\"\n",
-                "if [ -d build/packages ]; then state=present; else state=absent; fi\n",
-                "echo \"$(basename \"$PWD\") $1 $state\" >> \"$root/.fake/build-state\"\n",
-                "if [ \"$1\" = publish ]; then\n",
-                "  cp gleam.toml \"$root/.fake/published-$(basename \"$PWD\").toml\"\n",
-                "fi\n",
-            ),
-            root = root.display()
-        ),
-    );
-    fs::set_permissions(&script, fs::Permissions::from_mode(0o755)).unwrap();
-    fs::create_dir_all(root.join(".fake")).unwrap();
-    script
-}
-
-/// A trellis invocation aimed at the mock GitHub API: base URL and repo
-/// overridden, and a token in the environment.
-fn trellis_github(dir: &Path, api: &str) -> Command {
-    let mut cmd = trellis(dir);
-    cmd.env("TRELLIS_GITHUB_API_URL", api)
-        .env("TRELLIS_GITHUB_REPO", "example/repo")
-        .env("GITHUB_TOKEN", "test-token");
-    cmd
-}
-
-/// Serve a canned Hex API from a background thread: `versions` maps package
-/// name → published versions; unknown packages get a 404, like Hex.
-fn mock_hex(versions: Vec<(&'static str, Vec<&'static str>)>) -> String {
-    let listener = TcpListener::bind("127.0.0.1:0").unwrap();
-    let base = format!("http://{}/api", listener.local_addr().unwrap());
-    std::thread::spawn(move || {
-        for stream in listener.incoming() {
-            let Ok(mut stream) = stream else { continue };
-            let mut buf = [0u8; 4096];
-            let mut request = Vec::new();
-            // Read until end of headers (requests have no body).
-            loop {
-                match stream.read(&mut buf) {
-                    Ok(0) => break,
-                    Ok(n) => {
-                        request.extend_from_slice(&buf[..n]);
-                        if request.windows(4).any(|w| w == b"\r\n\r\n") {
-                            break;
-                        }
-                    }
-                    Err(_) => break,
-                }
-            }
-            let request = String::from_utf8_lossy(&request);
-            let path = request.split_whitespace().nth(1).unwrap_or("");
-            let package = path.rsplit('/').next().unwrap_or("");
-            let response = match versions.iter().find(|(name, _)| *name == package) {
-                Some((_, released)) => {
-                    let releases: Vec<String> = released
-                        .iter()
-                        .map(|v| format!("{{\"version\":\"{v}\"}}"))
-                        .collect();
-                    let body = format!("{{\"releases\":[{}]}}", releases.join(","));
-                    format!(
-                        "HTTP/1.1 200 OK\r\nContent-Type: application/json\r\nContent-Length: {}\r\nConnection: close\r\n\r\n{body}",
-                        body.len()
-                    )
-                }
-                None => "HTTP/1.1 404 Not Found\r\nContent-Length: 0\r\nConnection: close\r\n\r\n"
-                    .to_string(),
-            };
-            let _ = stream.write_all(response.as_bytes());
-        }
-    });
-    base
-}
 
 // ---- tag ----------------------------------------------------------------
 
@@ -390,297 +229,6 @@ fn tag_create_rejects_divergent_local_and_remote_tags() {
         .failure()
         .stderr(predicate::str::contains("different objects"));
 }
-
-// ---- ci tag-package -------------------------------------------------------
-
-#[test]
-fn ci_tag_package_resolves_tag_to_package() {
-    trellis(&fixture("basic"))
-        .args(["ci", "tag-package", "lat_core-v1.2.0"])
-        .assert()
-        .success()
-        .stdout("lat_core\n");
-
-    let output = trellis(&fixture("basic"))
-        .args(["ci", "tag-package", "lat_mid-v9.9.9", "--json"])
-        .output()
-        .unwrap();
-    assert!(output.status.success());
-    let info: serde_json::Value = serde_json::from_slice(&output.stdout).unwrap();
-    assert_eq!(info["name"], "lat_mid");
-    assert_eq!(info["version"], "0.5.0");
-    assert_eq!(info["tag_version"], "9.9.9");
-
-    trellis(&fixture("basic"))
-        .args(["ci", "tag-package", "unrelated-v1.0.0"])
-        .assert()
-        .failure()
-        .stderr(predicate::str::contains(
-            "does not match any releasable package",
-        ));
-}
-
-// ---- publish --------------------------------------------------------------
-
-#[test]
-fn publish_rewrites_path_deps_and_restores_the_manifest() {
-    let tmp = tempfile::tempdir().unwrap();
-    let root = tmp.path();
-    copy_fixture_to(root);
-    let gleam = install_fake_gleam(root);
-    let hex = mock_hex(vec![("lat_core", vec!["1.2.0"])]); // lat_mid: 404
-
-    let original = fs::read_to_string(root.join("packages/lat_mid/gleam.toml")).unwrap();
-    trellis(root)
-        .env("TRELLIS_GLEAM_BIN", &gleam)
-        .env("TRELLIS_HEX_API_URL", &hex)
-        .args(["publish", "lat_mid"])
-        .assert()
-        .success()
-        .stdout(predicate::str::contains(
-            "rewrote lat_core -> \">= 1.2.0 and < 2.0.0\"",
-        ))
-        .stdout(predicate::str::contains("[lat_mid] published 0.5.0"));
-
-    // Validation and publish ran, in order, in the package directory.
-    let log = fs::read_to_string(root.join(".fake/gleam-log")).unwrap();
-    assert_eq!(
-        log,
-        concat!(
-            "lat_mid gleam format --check\n",
-            "lat_mid gleam build --warnings-as-errors\n",
-            "lat_mid gleam test\n",
-            "lat_mid gleam publish --yes\n",
-        )
-    );
-    // gleam publish saw the rewritten manifest…
-    let published = fs::read_to_string(root.join(".fake/published-lat_mid.toml")).unwrap();
-    assert!(published.contains("lat_core = \">= 1.2.0 and < 2.0.0\""));
-    assert!(!published.contains("path"));
-    // …but the repo shows the original afterwards.
-    assert_eq!(
-        fs::read_to_string(root.join("packages/lat_mid/gleam.toml")).unwrap(),
-        original
-    );
-}
-
-#[test]
-fn publish_clears_the_stale_dependency_tree_before_publishing() {
-    let tmp = tempfile::tempdir().unwrap();
-    let root = tmp.path();
-    copy_fixture_to(root);
-    let gleam = install_fake_gleam(root);
-    let hex = mock_hex(vec![("lat_core", vec!["1.2.0"])]); // lat_mid: 404
-
-    // Stand in for what validation leaves behind: a dependency tree gleam
-    // resolved while lat_core was still a path dep, alongside compiled output.
-    let packages = root.join("packages/lat_mid/build/packages");
-    fs::create_dir_all(&packages).unwrap();
-    write(&packages.join("lat_core.config_fingerprint"), "stale");
-    let compiled = root.join("packages/lat_mid/build/dev/erlang/lat_mid");
-    fs::create_dir_all(&compiled).unwrap();
-    write(&compiled.join("lat_mid.app"), "compiled");
-
-    trellis(root)
-        .env("TRELLIS_GLEAM_BIN", &gleam)
-        .env("TRELLIS_HEX_API_URL", &hex)
-        .args(["publish", "lat_mid"])
-        .assert()
-        .success();
-
-    // Validation runs against the tree as resolved; publish must not, or gleam
-    // fails reading the local dependency it just dropped from the manifest.
-    let state = fs::read_to_string(root.join(".fake/build-state")).unwrap();
-    assert_eq!(
-        state,
-        concat!(
-            "lat_mid format present\n",
-            "lat_mid build present\n",
-            "lat_mid test present\n",
-            "lat_mid publish absent\n",
-        )
-    );
-    // Only the resolved-dependency half is stale; compiled output survives.
-    assert!(
-        compiled.join("lat_mid.app").exists(),
-        "clearing the dependency tree must not throw away compiled artifacts"
-    );
-}
-
-#[test]
-fn publish_keeps_the_dependency_tree_when_nothing_is_rewritten() {
-    let tmp = tempfile::tempdir().unwrap();
-    let root = tmp.path();
-    copy_fixture_to(root);
-    let gleam = install_fake_gleam(root);
-    let hex = mock_hex(vec![]); // nothing published yet
-
-    // lat_core has no path deps, so its manifest is published as-is and its
-    // dependency tree stays valid throughout — no reason to pay for a refetch.
-    let packages = root.join("packages/lat_core/build/packages");
-    fs::create_dir_all(&packages).unwrap();
-    write(&packages.join("gleam.lock"), "");
-
-    trellis(root)
-        .env("TRELLIS_GLEAM_BIN", &gleam)
-        .env("TRELLIS_HEX_API_URL", &hex)
-        .args(["publish", "lat_core"])
-        .assert()
-        .success();
-
-    let state = fs::read_to_string(root.join(".fake/build-state")).unwrap();
-    assert!(
-        state.contains("lat_core publish present"),
-        "unrewritten package should keep its dependency tree:\n{state}"
-    );
-    assert!(packages.join("gleam.lock").exists());
-}
-
-#[test]
-fn publish_skips_versions_already_on_hex() {
-    let tmp = tempfile::tempdir().unwrap();
-    let root = tmp.path();
-    copy_fixture_to(root);
-    let gleam = install_fake_gleam(root);
-    let hex = mock_hex(vec![("lat_core", vec!["1.1.0", "1.2.0"])]);
-
-    trellis(root)
-        .env("TRELLIS_GLEAM_BIN", &gleam)
-        .env("TRELLIS_HEX_API_URL", &hex)
-        .args(["publish", "lat_core"])
-        .assert()
-        .success()
-        .stdout(predicate::str::contains(
-            "1.2.0 is already on Hex; skipping",
-        ));
-    assert!(
-        !root.join(".fake/gleam-log").exists(),
-        "no gleam command should run"
-    );
-}
-
-#[test]
-fn publish_all_untagged_goes_in_topological_order_and_is_idempotent() {
-    let tmp = tempfile::tempdir().unwrap();
-    let root = tmp.path();
-    copy_fixture_to(root);
-    let gleam = install_fake_gleam(root);
-    // lat_core already published at the current version; the others aren't.
-    let hex = mock_hex(vec![("lat_core", vec!["1.2.0"])]);
-
-    trellis(root)
-        .env("TRELLIS_GLEAM_BIN", &gleam)
-        .env("TRELLIS_HEX_API_URL", &hex)
-        .args(["publish", "--all-untagged"])
-        .assert()
-        .success()
-        .stdout(predicate::str::contains(
-            "[lat_core] 1.2.0 is already on Hex",
-        ))
-        .stdout(predicate::str::contains("[lat_mid] published 0.5.0"))
-        .stdout(predicate::str::contains("[lat_cli] published 0.3.1"));
-
-    let log = fs::read_to_string(root.join(".fake/gleam-log")).unwrap();
-    let mid = log.find("lat_mid gleam publish").unwrap();
-    let cli = log.find("lat_cli gleam publish").unwrap();
-    assert!(
-        mid < cli,
-        "dependency must publish before dependent:\n{log}"
-    );
-    assert!(
-        !log.contains("package_a"),
-        "@release-excluded members never publish"
-    );
-}
-
-#[test]
-fn publish_dry_run_reports_without_running_gleam() {
-    let tmp = tempfile::tempdir().unwrap();
-    let root = tmp.path();
-    copy_fixture_to(root);
-    let gleam = install_fake_gleam(root);
-    let hex = mock_hex(vec![]);
-
-    trellis(root)
-        .env("TRELLIS_GLEAM_BIN", &gleam)
-        .env("TRELLIS_HEX_API_URL", &hex)
-        .args(["publish", "lat_mid", "--dry-run"])
-        .assert()
-        .success()
-        .stdout(predicate::str::contains("[lat_mid] would publish 0.5.0"))
-        .stdout(predicate::str::contains(
-            "lat_core -> \">= 1.2.0 and < 2.0.0\"",
-        ));
-    assert!(!root.join(".fake/gleam-log").exists());
-}
-
-#[test]
-fn publish_by_tag_refuses_version_mismatch() {
-    let tmp = tempfile::tempdir().unwrap();
-    let root = tmp.path();
-    copy_fixture_to(root);
-    let gleam = install_fake_gleam(root);
-    let hex = mock_hex(vec![]);
-
-    trellis(root)
-        .env("TRELLIS_GLEAM_BIN", &gleam)
-        .env("TRELLIS_HEX_API_URL", &hex)
-        .args(["publish", "--tag", "lat_core-v9.9.9"])
-        .assert()
-        .failure()
-        .stderr(predicate::str::contains("refusing to publish"));
-
-    trellis(root)
-        .env("TRELLIS_GLEAM_BIN", &gleam)
-        .env("TRELLIS_HEX_API_URL", &hex)
-        .args(["publish", "--tag", "lat_core-v1.2.0"])
-        .assert()
-        .success()
-        .stdout(predicate::str::contains("[lat_core] published 1.2.0"));
-}
-
-#[test]
-fn publish_restores_manifest_even_when_publish_fails() {
-    let tmp = tempfile::tempdir().unwrap();
-    let root = tmp.path();
-    copy_fixture_to(root);
-    // gleam fails only on `publish`, after validation passes.
-    let gleam = root.join("fake-gleam.sh");
-    write(
-        &gleam,
-        "#!/bin/sh\nif [ \"$1\" = publish ]; then echo 'rate limited' >&2; exit 1; fi\n",
-    );
-    fs::set_permissions(&gleam, fs::Permissions::from_mode(0o755)).unwrap();
-    let hex = mock_hex(vec![]);
-
-    let original = fs::read_to_string(root.join("packages/lat_mid/gleam.toml")).unwrap();
-    trellis(root)
-        .env("TRELLIS_GLEAM_BIN", &gleam)
-        .env("TRELLIS_HEX_API_URL", &hex)
-        .args(["publish", "lat_mid"])
-        .assert()
-        .failure();
-    assert_eq!(
-        fs::read_to_string(root.join("packages/lat_mid/gleam.toml")).unwrap(),
-        original,
-        "gleam.toml must be restored after a failed publish"
-    );
-}
-
-#[test]
-fn publish_rejects_unreleasable_package() {
-    let tmp = tempfile::tempdir().unwrap();
-    let root = tmp.path();
-    copy_fixture_to(root);
-    trellis(root)
-        .args(["publish", "package_a"])
-        .assert()
-        .failure()
-        .stderr(predicate::str::contains(
-            "release lifecycle `workspace`, not `hex`",
-        ));
-}
-
 // ---- series tags -----------------------------------------------------------
 
 /// The repository-series config the tag tests share: `lat_cli` anchoring
@@ -690,68 +238,6 @@ const REPO_SERIES: &str = "repository_tag_package = \"lat_cli\"\n\
 
 /// A bare repository added to `root`'s repo as `origin`. The returned remote
 /// must outlive the test.
-fn bare_origin(root: &Path) -> tempfile::TempDir {
-    let remote = tempfile::tempdir().unwrap();
-    git(remote.path(), &["init", "-q", "--bare"]);
-    git(
-        root,
-        &["remote", "add", "origin", remote.path().to_str().unwrap()],
-    );
-    remote
-}
-
-/// The basic fixture with `[tools.trellis.publish]` replaced by `publish`, in
-/// a git repo with a bare origin. The returned remote must outlive the test.
-fn series_repo(root: &Path, publish: &str) -> tempfile::TempDir {
-    copy_fixture_to(root);
-    write(
-        &root.join("gleam.toml"),
-        &format!(
-            "[tools.trellis]\nmembers = [\"packages/*\", \"examples/*\"]\n\
-             exclude = {{ \"@release\" = [\"examples/*\"] }}\n\n\
-             [tools.trellis.publish]\n{publish}\n"
-        ),
-    );
-    init_repo(root);
-    bare_origin(root)
-}
-
-fn set_version(root: &Path, package: &str, version: &str) {
-    let path = root.join("packages").join(package).join("gleam.toml");
-    let text = fs::read_to_string(&path).unwrap();
-    let text: Vec<String> = text
-        .lines()
-        .map(|line| {
-            if line.starts_with("version = ") {
-                format!("version = \"{version}\"")
-            } else {
-                line.to_string()
-            }
-        })
-        .collect();
-    fs::write(&path, text.join("\n") + "\n").unwrap();
-}
-
-fn git_stdout(dir: &Path, args: &[&str]) -> String {
-    let output = std::process::Command::new("git")
-        .args(["-c", "safe.bareRepository=all"])
-        .args(args)
-        .current_dir(dir)
-        .output()
-        .unwrap();
-    assert!(
-        output.status.success(),
-        "git {args:?} failed in {}: {}",
-        dir.display(),
-        String::from_utf8_lossy(&output.stderr)
-    );
-    String::from_utf8_lossy(&output.stdout).trim().to_string()
-}
-
-fn commit_of(dir: &Path, revision: &str) -> String {
-    git_stdout(dir, &["rev-parse", &format!("{revision}^{{commit}}")])
-}
-
 #[test]
 fn series_tag_moves_with_each_release_while_exact_tags_stay_put() {
     let tmp = tempfile::tempdir().unwrap();
@@ -815,29 +301,6 @@ fn series_tag_moves_with_each_release_while_exact_tags_stay_put() {
         .assert()
         .success()
         .stdout(predicate::str::contains("already tagged"));
-}
-
-#[test]
-fn info_reports_the_tags_a_package_actually_gets() {
-    let tmp = tempfile::tempdir().unwrap();
-    let root = tmp.path();
-    let _remote = series_repo(
-        root,
-        "package_tags_overrides = { \"packages/lat_cli\" = [\"exact\", \"minor\"] }",
-    );
-
-    trellis(root)
-        .args(["info", "lat_cli"])
-        .assert()
-        .success()
-        .stdout(predicate::str::contains("tag:        lat_cli-v0.3.1"))
-        .stdout(predicate::str::contains("series tag: lat_cli-v0.3"));
-    trellis(root)
-        .args(["info", "lat_core"])
-        .assert()
-        .success()
-        .stdout(predicate::str::contains("tag:        lat_core-v1.2.0"))
-        .stdout(predicate::str::contains("series tag:").not());
 }
 
 #[test]
@@ -1534,270 +997,4 @@ fn doctor_rejects_a_package_tags_override_that_matches_nothing() {
         .stdout(predicate::str::contains(
             "`package_tags_overrides` glob `packages/nope` matches no member",
         ));
-}
-
-// ---- release bootstrap -----------------------------------------------------
-
-// Mirrors `version_of` in phase2.rs — the same manifest-version read, kept in
-// step until the test binaries grow a shared support module.
-fn version_of(root: &Path, package: &str) -> String {
-    let manifest = fs::read_to_string(root.join("packages").join(package).join("gleam.toml"))
-        .unwrap_or_else(|err| panic!("no gleam.toml for {package}: {err}"));
-    manifest
-        .lines()
-        .find_map(|line| line.trim().strip_prefix("version = "))
-        .unwrap_or_else(|| panic!("no version in {package}'s gleam.toml"))
-        .trim_matches('"')
-        .to_string()
-}
-
-#[test]
-fn bootstrap_uses_current_versions_with_no_fragments_required() {
-    let tmp = tempfile::tempdir().unwrap();
-    let root = tmp.path();
-    copy_fixture_to(root);
-    init_repo(root);
-    // The fixture ships no `.changes/unreleased` directory at all — bootstrap
-    // must not need one.
-    assert!(!root.join(".changes").exists());
-
-    trellis(root)
-        .args(["release", "bootstrap"])
-        .assert()
-        .success()
-        .stdout(predicate::str::contains("tagged lat_core-v1.2.0"))
-        .stdout(predicate::str::contains("tagged lat_mid-v0.5.0"))
-        .stdout(predicate::str::contains("tagged lat_cli-v0.3.1"));
-
-    let tags = git_stdout(root, &["tag", "--list"]);
-    assert_eq!(
-        tags.lines().collect::<Vec<_>>(),
-        vec!["lat_cli-v0.3.1", "lat_core-v1.2.0", "lat_mid-v0.5.0"]
-    );
-    // No version bump: bootstrap never runs `version plan`/`version apply`.
-    assert_eq!(version_of(root, "lat_core"), "1.2.0");
-    assert_eq!(version_of(root, "lat_mid"), "0.5.0");
-    assert_eq!(version_of(root, "lat_cli"), "0.3.1");
-}
-
-#[test]
-fn bootstrap_dry_run_reports_every_action_and_mutates_nothing() {
-    let tmp = tempfile::tempdir().unwrap();
-    let root = tmp.path();
-    copy_fixture_to(root);
-    init_repo(root);
-    let remote = bare_origin(root);
-    let api = common::mock_github(root);
-
-    trellis_github(root, &api)
-        .args([
-            "release",
-            "bootstrap",
-            "--dry-run",
-            "--push",
-            "--github-release",
-        ])
-        .assert()
-        .success()
-        .stdout(predicate::str::contains("would tag lat_core-v1.2.0"))
-        .stdout(predicate::str::contains("would push lat_core-v1.2.0"))
-        .stdout(predicate::str::contains(
-            "would create GitHub release lat_core-v1.2.0",
-        ))
-        .stdout(predicate::str::contains("would tag lat_mid-v0.5.0"))
-        .stdout(predicate::str::contains(
-            "would create GitHub release lat_cli-v0.3.1",
-        ));
-
-    // Nothing was actually created, locally, on origin, or via the API.
-    assert_eq!(git_stdout(root, &["tag", "--list"]), "");
-    assert_eq!(git_stdout(remote.path(), &["tag", "--list"]), "");
-    let log = fs::read_to_string(root.join(".fake/github-log")).unwrap_or_default();
-    assert!(
-        !log.contains("POST /repos/example/repo/releases\n"),
-        "{log}"
-    );
-}
-
-#[test]
-fn bootstrap_creates_both_exact_and_series_tags() {
-    let tmp = tempfile::tempdir().unwrap();
-    let root = tmp.path();
-    let _remote = series_repo(
-        root,
-        "package_tags_overrides = { \"packages/lat_cli\" = [\"exact\", \"minor\"] }",
-    );
-
-    trellis(root)
-        .args(["release", "bootstrap", "--dry-run"])
-        .assert()
-        .success()
-        .stdout(predicate::str::contains("would tag lat_cli-v0.3.1"))
-        .stdout(predicate::str::contains("would tag lat_cli-v0.3\n"));
-
-    trellis(root)
-        .args(["release", "bootstrap"])
-        .assert()
-        .success();
-    let tags = git_stdout(root, &["tag", "--list"]);
-    assert_eq!(
-        tags.lines().collect::<Vec<_>>(),
-        vec![
-            "lat_cli-v0.3",
-            "lat_cli-v0.3.1",
-            "lat_core-v1.2.0",
-            "lat_mid-v0.5.0"
-        ]
-    );
-}
-
-#[test]
-fn bootstrap_leaves_release_excluded_packages_untagged() {
-    let tmp = tempfile::tempdir().unwrap();
-    let root = tmp.path();
-    copy_fixture_to(root);
-    init_repo(root);
-
-    trellis(root)
-        .args(["release", "bootstrap", "--dry-run"])
-        .assert()
-        .success()
-        .stdout(predicate::str::contains("package_a").not());
-
-    trellis(root)
-        .args(["release", "bootstrap"])
-        .assert()
-        .success();
-    let tags = git_stdout(root, &["tag", "--list"]);
-    assert!(!tags.contains("package_a"), "{tags}");
-}
-
-#[test]
-fn bootstrap_fetches_a_remote_only_tag_instead_of_recreating_it() {
-    let tmp = tempfile::tempdir().unwrap();
-    let root = tmp.path();
-    copy_fixture_to(root);
-    init_repo(root);
-    let _remote = bare_origin(root);
-    // lat_core is already tagged and pushed by some earlier process; the
-    // local clone bootstrapping now has never seen the tag.
-    git(root, &["tag", "-a", "lat_core-v1.2.0", "-m", "existing"]);
-    git(root, &["push", "origin", "lat_core-v1.2.0"]);
-    let original = commit_of(root, "lat_core-v1.2.0");
-    git(root, &["tag", "-d", "lat_core-v1.2.0"]);
-
-    trellis(root)
-        .args(["release", "bootstrap", "--dry-run", "--push"])
-        .assert()
-        .success()
-        .stdout(predicate::str::contains("would fetch lat_core-v1.2.0"));
-
-    trellis(root)
-        .args(["release", "bootstrap", "--push"])
-        .assert()
-        .success()
-        .stdout(predicate::str::contains("fetched lat_core-v1.2.0"));
-    assert_eq!(commit_of(root, "lat_core-v1.2.0"), original);
-}
-
-#[test]
-fn bootstrap_reports_existing_releases_and_reruns_idempotently() {
-    let tmp = tempfile::tempdir().unwrap();
-    let root = tmp.path();
-    copy_fixture_to(root);
-    init_repo(root);
-    let _remote = bare_origin(root);
-    let api = common::mock_github(root);
-
-    trellis_github(root, &api)
-        .args(["release", "bootstrap", "--push", "--github-release"])
-        .assert()
-        .success();
-    let first_log = fs::read_to_string(root.join(".fake/github-log")).unwrap();
-    assert_eq!(
-        first_log
-            .matches("POST /repos/example/repo/releases\n")
-            .count(),
-        3
-    );
-
-    trellis_github(root, &api)
-        .args([
-            "release",
-            "bootstrap",
-            "--dry-run",
-            "--push",
-            "--github-release",
-        ])
-        .assert()
-        .success()
-        .stdout(predicate::str::contains(
-            "GitHub release lat_core-v1.2.0 already exists; skipping",
-        ))
-        .stdout(predicate::str::contains("would").not());
-
-    trellis_github(root, &api)
-        .args(["release", "bootstrap", "--push", "--github-release"])
-        .assert()
-        .success();
-    let second_log = fs::read_to_string(root.join(".fake/github-log")).unwrap();
-    assert_eq!(
-        second_log
-            .matches("POST /repos/example/repo/releases\n")
-            .count(),
-        3
-    );
-}
-
-#[test]
-fn bootstrap_preflights_conflicts_before_mutating_any_package() {
-    let tmp = tempfile::tempdir().unwrap();
-    let root = tmp.path();
-    copy_fixture_to(root);
-    init_repo(root);
-    let _remote = bare_origin(root);
-    // lat_core is tagged and pushed, then the local tag is force-moved to a
-    // later commit — an immutable tag disagreeing with origin.
-    git(root, &["tag", "-a", "lat_core-v1.2.0", "-m", "first"]);
-    git(root, &["push", "origin", "lat_core-v1.2.0"]);
-    write(&root.join("later.txt"), "later\n");
-    git(root, &["add", "."]);
-    git(root, &["commit", "-q", "-m", "later"]);
-    git(root, &["tag", "-f", "-a", "lat_core-v1.2.0", "-m", "moved"]);
-    // lat_mid and lat_cli have no tag yet, and would otherwise be created.
-
-    trellis(root)
-        .args(["release", "bootstrap", "--dry-run", "--push"])
-        .assert()
-        .failure()
-        .stderr(predicate::str::contains("different objects"));
-
-    trellis(root)
-        .args(["release", "bootstrap", "--push"])
-        .assert()
-        .failure()
-        .stderr(predicate::str::contains("different objects"));
-
-    // Neither the dry-run nor the failed real run tagged anything else —
-    // the conflict on lat_core blocks the whole batch.
-    let tags = git_stdout(root, &["tag", "--list"]);
-    assert_eq!(tags.lines().collect::<Vec<_>>(), vec!["lat_core-v1.2.0"]);
-}
-
-// ---- lockfile refresh ------------------------------------------------------
-
-#[test]
-fn lockfile_refresh_scopes_to_one_package() {
-    let tmp = tempfile::tempdir().unwrap();
-    let root = tmp.path();
-    copy_fixture_to(root);
-    let gleam = install_fake_gleam(root);
-
-    trellis(root)
-        .env("TRELLIS_GLEAM_BIN", &gleam)
-        .args(["lockfile", "refresh", "--package", "lat_mid"])
-        .assert()
-        .success();
-    let log = fs::read_to_string(root.join(".fake/gleam-log")).unwrap();
-    assert_eq!(log, "lat_mid gleam deps download\n");
 }
